refactor: Move error parsers to their own module
