+++ src/combinator.rs
@@ -2,8 +2,7 @@ use lib::marker::PhantomData;
 use lib::mem;
 
 use Parser;
-use error::{Consumed, ConsumedResult, Info, ParseError, ParseResult, StreamError, Tracked,
-            UnexpectedParse};
+use error::{Consumed, ConsumedResult, Info, ParseError, ParseResult, Tracked};
 use stream::{Positioned, Resetable, Stream, StreamOnce};
 use parser::ParseMode;
 
@@ -19,54 +18,8 @@ pub use parser::choice::*;
 pub use parser::item::*;
 #[doc(inline)]
 pub use parser::repeat::*;
-
-#[derive(Clone)]
-pub struct Unexpected<I>(Info<I::Item, I::Range>, PhantomData<fn(I) -> I>)
-where
-    I: Stream;
-impl<I> Parser for Unexpected<I>
-where
-    I: Stream,
-{
-    type Input = I;
-    type Output = ();
-    type PartialState = ();
-    #[inline]
-    fn parse_lazy(&mut self, input: &mut Self::Input) -> ConsumedResult<(), I> {
-        EmptyErr(<Self::Input as StreamOnce>::Error::empty(input.position()).into())
-    }
-    fn add_error(&mut self, errors: &mut Tracked<<Self::Input as StreamOnce>::Error>) {
-        errors.error.add(StreamError::unexpected(self.0.clone()));
-    }
-}
-/// Always fails with `message` as an unexpected error.
-/// Never consumes any input.
-///
-/// ```
-/// # extern crate combine;
-/// # use combine::*;
-/// # use combine::error::StreamError;
-/// # fn main() {
-/// let result = unexpected("token")
-///     .easy_parse("a");
-/// assert!(result.is_err());
-/// assert!(
-///     result.err()
-///         .unwrap()
-///         .errors
-///         .iter()
-///         .any(|m| *m == StreamError::unexpected("token".into()))
-/// );
-/// # }
-/// ```
-#[inline(always)]
-pub fn unexpected<I, S>(message: S) -> Unexpected<I>
-where
-    I: Stream,
-    S: Into<Info<I::Item, I::Range>>,
-{
-    Unexpected(message.into(), PhantomData)
-}
+#[doc(inline)]
+pub use parser::error::*;
 
 parser!{
 #[derive(Copy, Clone)]
@@ -277,68 +230,6 @@ where
     LookAhead(p)
 }
 
-#[derive(Clone)]
-pub struct Message<P>(
-    P,
-    Info<<P::Input as StreamOnce>::Item, <P::Input as StreamOnce>::Range>,
-)
-where
-    P: Parser;
-impl<I, P> Parser for Message<P>
-where
-    I: Stream,
-    P: Parser<Input = I>,
-{
-    type Input = I;
-    type Output = P::Output;
-    type PartialState = P::PartialState;
-
-    parse_mode!();
-    #[inline]
-    fn parse_mode_impl<M>(
-        &mut self,
-        mode: M,
-        input: &mut Self::Input,
-        state: &mut Self::PartialState,
-    ) -> ConsumedResult<Self::Output, Self::Input>
-    where
-        M: ParseMode,
-    {
-        match self.0.parse_mode(mode, input, state) {
-            ConsumedOk(x) => ConsumedOk(x),
-            EmptyOk(x) => EmptyOk(x),
-
-            // The message should always be added even if some input was consumed before failing
-            ConsumedErr(mut err) => {
-                err.add_message(self.1.clone());
-                ConsumedErr(err)
-            }
-
-            // The message will be added in `add_error`
-            EmptyErr(err) => EmptyErr(err),
-        }
-    }
-
-    fn add_error(&mut self, errors: &mut Tracked<<Self::Input as StreamOnce>::Error>) {
-        self.0.add_error(errors);
-        errors.error.add_message(self.1.clone());
-    }
-}
-
-/// Equivalent to [`p1.message(msg)`].
-///
-/// [`p1.message(msg)`]: ../error/trait.Parser.html#method.message
-#[inline(always)]
-pub fn message<P>(
-    p: P,
-    msg: Info<<P::Input as StreamOnce>::Item, <P::Input as StreamOnce>::Range>,
-) -> Message<P>
-where
-    P: Parser,
-{
-    Message(p, msg)
-}
-
 #[derive(Copy, Clone)]
 pub struct Map<P, F>(P, F)
 where
@@ -441,56 +332,6 @@ where
     F: FnMut(P::Output) -> Result<B, <P::Input as StreamOnce>::Error>,
 {
     FlatMap(p, f)
-}
-
-#[derive(Clone)]
-pub struct Expected<P>(
-    P,
-    Info<<P::Input as StreamOnce>::Item, <P::Input as StreamOnce>::Range>,
-)
-where
-    P: Parser;
-impl<P> Parser for Expected<P>
-where
-    P: Parser,
-{
-    type Input = P::Input;
-    type Output = P::Output;
-    type PartialState = P::PartialState;
-
-    parse_mode!();
-    #[inline(always)]
-    fn parse_mode_impl<M>(
-        &mut self,
-        mode: M,
-        input: &mut Self::Input,
-        state: &mut Self::PartialState,
-    ) -> ConsumedResult<Self::Output, Self::Input>
-    where
-        M: ParseMode,
-    {
-        self.0.parse_mode(mode, input, state)
-    }
-
-    fn add_error(&mut self, errors: &mut Tracked<<Self::Input as StreamOnce>::Error>) {
-        ParseError::set_expected(errors, StreamError::expected(self.1.clone()), |errors| {
-            self.0.add_error(errors);
-        })
-    }
-}
-
-/// Equivalent to [`p.expected(info)`].
-///
-/// [`p.expected(info)`]: ../error/trait.Parser.html#method.expected
-#[inline(always)]
-pub fn expected<P>(
-    p: P,
-    info: Info<<P::Input as StreamOnce>::Item, <P::Input as StreamOnce>::Range>,
-) -> Expected<P>
-where
-    P: Parser,
-{
-    Expected(p, info)
 }
 
 #[derive(Copy, Clone)]
