+++ src/parser/mod.rs
@@ -74,6 +74,8 @@ pub mod byte;
 pub mod char;
 /// Parsers working with single stream items.
 pub mod item;
+/// Parsers which cause errors or modifies the returned error on parse failure.
+pub mod error;
 /// Combinators which take one or more parsers and apply them repeatedly.
 pub mod repeat;
 pub mod sequence;
